requirements.txt
deta==1.1.0
plotly==5.11.0
python-dotenv==0.21.0
streamlit==1.16.0
streamlit_option_menu==0.3.2
